requirements.txt
fasthtml==0.0.0a0
google-generativeai==0.8.3
Pillow==10.4.0
PyMuPDF==1.24.3
